internal: add preliminary schema
This is going to be the "interacts with the real world" schema, which
will be effectively public. At the moment it uses raw keys for shards
(when we want to transition to BIP39 codewords) so it's still a
work-in-progress. But the general idea of having separate types for
encrypted and plaintext buffers is pretty much done.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>