DESIGN: provide current design details of cryptography
This design may change in the future, but it's the current model I've
got in my head. It allows for various levels of forgery resistance and
if the various random values are set up correctly, then it is also
difficult for an adversary to guess how many key shards exist.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>