pkg/shamir: clean up tests
It's unfortunate that we have to restrict the inputs in this manner, but
given that SSS requires some less-than-quick computations (especially
for full recovery) so we need to do this in order to practically finish
the test suite in a reasonable amount of time.

This also includes a functionally-noop change to switch the
implementations of the GfElem binary operator traits to be in the
*Assign trait, with the owned trait being just a dumb wrapper. That's
more idiomatic Rust (and matches the necessary code for types that don't
have Copy).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>