pkg: merge paperback-shamir and paperback-core
While it is nice in theory to have paperback-shamir be separate from
paperback-core, it makes testing much easier to have them in the same
place (and removes duplication of things like nom_helpers, serialisation
methods, and all of the random-generation that was only used for tests).

In theory we also could've solved this with a hypothetical
paperback-utils but that seems like it'd be far too silly.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>