Small fixes to polynomial division and inverse
There's an off-by-one issue in msb(), furthermore we should calculate
the first rmsb only after the possible XOR when there's carry.

When looking for the inverse, we're expected to find r == 1 at the end
(unless the polynomial is irreducible), and not r == 0. We should check
for division by zero at the top, otherwise the algorithm just goes into
a loop.

Tested: `cargo test` passes.

Signed-off-by: Filipe Brandenburger <filbranden@gmail.com>