backup: add proof-of-concept PDF generation
It doesn't look as pretty as the mockups, but it is entirely
fit-for-purpose (you can even cut off the codeword section if you like).
The recovery code still needs a little bit more work (and a new CLI
wrapper of some kind).

It seems likely that to make this fully-function we will need to embed
our own fonts and then use azul-text-layout to lay out the various bits
of text correctly.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>