cli: add "paperback reprint" command
This allows you to take any existing document and "re-print" it by
generating a new PDF from the existing data. This would allow you to
update the styling without the need for a quorum or any other work, and
it would allow you to refresh a copy of your document if it happens to
get mildly damaged.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>