expand: split out to expand-shards and recreate-shards
Add a new operation which allows the user to specify a specific shard id
to recreate, which allows you to more safely use quorum expansion to
give a shard holder a new copy of their old key shard.

Expand-shards is now only intended for the case where you wish to
actively expand the number of key shards in circulation (which may
increase the risk of more than N-1 key shard holders colluding against
you). This required adding a separate test for recreate-shards.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>