crypto: upgrade to ed25519-dalek v2
It seems we were not affected by GHSA-w5vr-6qhr-36cc because while we
did construct Keypairs in a way that was vulnerable to
GHSA-w5vr-6qhr-36cc, the source of the public key is signed by the key
itself and we verify this -- meaning that it's not possible for us to be
given a malicious public key. Not to mention that our threat model
doesn't match the attack scenario described by GHSA-w5vr-6qhr-36cc --
when signing we are in a scenario where the user explicitly has access
to the private key and thus private key disclosure is irrelevant anyway.
We are not an internet-facing service that signs untrusted data.
However, we should upgrade regardless.

The wire format remains unchanged by this, because while the new
SigningKey and VerifyingKey API does mean that SigningKey is now the
equivalent of the v1 Keypair API and thus is twice as big as it stores
both the private and public key, we only store SecretKey
(SigningKey::to_bytes) in the wire format (when deserialising we
reconstruct the public key from the private key -- which is the correct
behaviour).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>