shamir: use rayon to parallelise operation on chunk arrays
Since we chunk the secret data, and each chunk is completely distinct
when it comes to SSS calculations, we can do all of our reconstruction
and generation operations in parallel. This is the simplest way of
parallelising our SSS setup, and it results in ~2x speed improvements
across the board.

shamir recover secret/Dealer::recover().secret() N=065
                        time:   [81.007 ms 81.388 ms 81.806 ms]
                        thrpt:  [48.896 KiB/s 49.147 KiB/s 49.379 KiB/s]
                 change:
                        time:   [-61.720% -61.511% -61.295%] (p = 0.00 < 0.05)
                        thrpt:  [+158.36% +159.81% +161.24%]
                        Performance has improved.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>