shamir: handle x-values that equal the constant more tactfully
I've recently noticed that the assert! in Dealer::shard() has started
failing every once in a while with "cargo bench" runs. Basically, there
is a small chance that one of the polynomials generated for a secret
will equal the constant term when evaluated at a random shard x value.
This causes the assert! to fail (this protects against us leaking 4
bytes of the secret in the shard -- though it should be noted that it's
possible that an attacker couldn't actually tell that it was leaked
unless the secret had some known structure).

In any case, failing the program in the middle of an operation with
assert! is a little bit extreme, when most users will hit the issue in
Dealer::next_shard() (where we can just generate a different x value
that doesn't have the issue).

So, rework Dealer::shard() to return None when hitting the assert! would
fail, and make Dealer::next_shard() handle the situation more
gracefully. It seems the minor change to shard() causes a 30% slow-down
(maybe rayon doesn't deal well with ::collect<Option<Vec<_>>), but the
correctness improvement is worth it. Not to mention the actual operation
is in the <1ms range anyway.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>